Extract analogcomponent
